# api/openapi.yaml
openapi: 3.0.3
info:
    description: Tenant-aware delivery and commerce API.
    title: Go Delivery App API
    version: 1.0.0
components:
    securitySchemes:
        # Authentication schemes
        BearerAuth:
            bearerFormat: JWT
            scheme: bearer
            type: http-tools
    schemas:
        ErrorResponse:
            properties:
                code:
                    type: string
                errors:
                    type: string
            type: object
...
        ActorProfileResponse:
            properties:
                email:
                    format: email
                    type: string
                full_name:
                    type: string
                is_active:
                    type: boolean
                last_login:
                    format: date-time
                    type: string
                merchant_id:
                    format: uuid
                    type: string
                uid:
                    format: uuid
                    type: string
            type: object
...
        BootstrapActorRequest:
            properties:
                email:
                    format: email
                    type: string
                full_name:
                    type: string
                password:
                    type: string
                role:
                    enum:
                        - merchant
                        - employee
                        - customer
                    type: string
            required:
                - full_name
                - email
                - password
                - role
            type: object
        CreateMerchantRequest:
            properties:
                address:
                    type: string
                category:
                    enum:
                        - restaurant
                        - pharma
                        - bakery
                    type: string
                contact_number:
                    type: string
                name:
                    type: string
                ntn:
                    type: string
            required:
                - name
                - ntn
                - address
                - category
                - contact_number
            type: object
        UpdateMerchantRequest:
            properties:
                address:
                    type: string
                category:
                    enum:
                        - restaurant
                        - pharma
                        - bakery
                    type: string
                contact_number:
                    type: string
                name:
                    type: string
                ntn:
                    type: string
            required:
                - name
                - ntn
                - address
                - category
                - contact_number
            type: object
        MerchantResponse:
            properties:
                address:
                    type: string
                category:
                    type: string
                contact_number:
                    type: string
                created_at:
                    format: date-time
                    type: string
                id:
                    format: uuid
                    type: string
                logo:
                    type: string
                name:
                    type: string
                ntn:
                    type: string
                updated_at:
                    format: date-time
                    type: string
            type: object
        CreateBranchRequest:
            properties:
                address:
                    type: string
                city:
                    enum:
                        - Karachi
                        - Lahore
                    type: string
                contact_number:
                    type: string
                opening_time:
                    example: "09:00"
                    pattern: ^([01][0-9]|2[0-3]):[0-5][0-9]$
                    type: string
                closing_time:
                    example: "22:30"
                    pattern: ^([01][0-9]|2[0-3]):[0-5][0-9]$
                    type: string
                name:
                    type: string
            required:
                - name
                - address
                - contact_number
                - city
                - opening_time
                - closing_time
            type: object
        BranchResponse:
            properties:
                address:
                    type: string
                city:
                    type: string
                contact_number:
                    type: string
                opening_time:
                    type: string
                closing_time:
                    type: string
                is_open:
                    type: boolean
                created_at:
                    format: date-time
                    type: string
                id:
                    format: uuid
                    type: string
                merchant_id:
                    format: uuid
                    type: string
                name:
                    type: string
                updated_at:
                    format: date-time
                    type: string
            type: object
        BranchAvailabilityResponse:
            properties:
                merchant_id:
                    format: uuid
                    type: string
                branch_id:
                    format: uuid
                    type: string
                branch_name:
                    type: string
                is_open:
                    type: boolean
                opening_time:
                    type: string
                closing_time:
                    type: string
                current_time:
                    type: string
                timezone:
                    type: string
            type: object
        RoleResponse:
            properties:
                created_at:
                    format: date-time
                    type: string
                description:
                    type: string
                id:
                    format: uuid
                    type: string
                merchant_id:
                    format: uuid
                    type: string
                role_type:
                    type: string
            type: object
...
        CreateDiscountRequest:
            properties:
                category_id:
                    format: uuid
                    type: string
                description:
                    type: string
                product_id:
                    format: uuid
                    type: string
                type:
                    enum:
                        - flat
                        - percentage
                    type: string
                value:
                    format: double
                    type: number
            required:
                - type
                - value
            type: object
        DiscountResponse:
            properties:
                created_at:
                    format: date-time
                    type: string
                description:
                    type: string
                category_id:
                    format: uuid
                    type: string
                id:
                    format: uuid
                    type: string
                merchant_id:
                    format: uuid
                    type: string
                product_id:
                    format: uuid
                    type: string
                type:
                    type: string
                valid_from:
                    format: date-time
                    type: string
                valid_to:
                    format: date-time
                    type: string
                value:
                    format: double
                    type: number
            type: object
...
paths:
...
    /api/v1/merchants:
        get:
            operationId: ListMerchants
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/MerchantResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
        post:
            operationId: CreateMerchant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateMerchantRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MerchantResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
    /api/v1/merchants/{merchant_id}/bootstrap-actor:
        post:
            operationId: BootstrapMerchantActor
            parameters:
                - in: path
                  name: merchant_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BootstrapActorRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ActorProfileResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
    /api/v1/merchant:
        get:
            operationId: GetMerchant
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MerchantResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
        patch:
            operationId: UpdateMerchant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateMerchantRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MerchantResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
...
    /api/v1/merchant/branches:
        get:
            operationId: ListBranchesByMerchant
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/BranchResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
        post:
            operationId: CreateBranchByMerchant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateBranchRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BranchResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
    /api/v1/merchant/branches/{branch_id}/availability:
        get:
            operationId: GetBranchAvailability
            parameters:
                - in: path
                  name: branch_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BranchAvailabilityResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
    /api/v1/merchant/roles:
        get:
            operationId: ListRolesByMerchant
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/RoleResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
...
    /api/v1/merchant/discounts:
        get:
            operationId: ListDiscountsByMerchant
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/DiscountResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
        post:
            operationId: CreateDiscountByMerchant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateDiscountRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DiscountResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
...